+++ packages/solid/CHANGELOG.md
@@ -11,6 +11,7 @@
   input field multiple times
 - **Presence**: Fixed a bug where elements get stuck in unmountSuspended state during rapid state updates
 - **Radio Group, Tabs**: Fixed indicator prematurely showing when rect has not been resolved yet
+- **ScrollArea**: Removed unnecessary `createMemo` wrapper in scrollbar props as is already reactive.
 - **Tabs**: Fixed tabs indicator position not updating when inactive tabs change size
 - **Tags Input**: Fixed issue where item delete trigger doesn't have `data-*` attached